fix_ftoa_1

+++ SandBox/ATMEGA128/Atmega128/Libraria/function.h
@@ -53,7 +53,7 @@ struct fnctn{
 	long (*gcd1)(long a, long b);
 	uint8_t (*pincheck)(uint8_t port, uint8_t pin);
 	char* (*print_binary)(uint8_t number);
-	char* (*ftoa)(float n, char* res, int afterpoint);
+	char* (*ftoa)(float n, char* res, uint8_t afterpoint);
 	/***pc use***
 	char* (*fltos)(FILE* stream);
 	char* (*ftos)(FILE* stream);
